Moved SubComponents from components subfolder to Nav component folder
